:ok_hand: split load waitPageactivity and UA waitpageactivity
